feat: security endpoints

+++ packages/ash-qrb-server/src/core/module/crud/crud.generate.ts
@@ -5,12 +5,24 @@ import type { XataHttpDatabase } from 'drizzle-orm/xata-http/driver'
 import { createInsertSchema, createSelectSchema } from 'drizzle-typebox'
 import { Elysia, t } from 'elysia'
 
+type CrudParametersIncludeAndExclude =
+  | 'getAll'
+  | 'getById'
+  | 'add'
+  | 'deleteById'
+  | 'patchById'
+export interface CrudParameters {
+  schema: AnyPgTable
+  include?: CrudParametersIncludeAndExclude[]
+  exclude?: CrudParametersIncludeAndExclude[]
+  security?: boolean
+}
 export class CrudApi {
   readonly db: XataHttpDatabase<Record<string, unknown>>
-  readonly schemaList: AnyPgTable[]
+  readonly schemaList: CrudParameters[]
   constructor(
     db: XataHttpDatabase<Record<string, unknown>>,
-    schemaList: AnyPgTable[],
+    schemaList: CrudParameters[],
   ) {
     this.db = db
     this.schemaList = schemaList
@@ -20,15 +32,32 @@ export class CrudApi {
     const elysia = new Elysia()
     elysia.decorate('plugin', 'crud')
 
-    for (const schema of this.schemaList) {
+    for (const crudParams of this.schemaList) {
+      const { schema, include, exclude } = crudParams
       // @ts-ignore
       const name = schema[Object.getOwnPropertySymbols(schema)[0]]
+      const condition = (name: CrudParametersIncludeAndExclude) =>
+        include?.includes(name) || !exclude?.includes(name)
 
-      this.getAll(name, elysia, schema)
-      this.getById(name, elysia, schema)
-      this.add(name, elysia, schema)
-      this.deleteById(name, elysia, schema)
-      this.patchById(name, elysia, schema)
+      if (condition('getAll')) {
+        this.getAll(name, elysia, schema)
+      }
+
+      if (condition('getById')) {
+        this.getById(name, elysia, schema)
+      }
+
+      if (condition('add')) {
+        this.add(name, elysia, schema)
+      }
+
+      if (condition('deleteById')) {
+        this.deleteById(name, elysia, schema)
+      }
+
+      if (condition('patchById')) {
+        this.patchById(name, elysia, schema)
+      }
     }
 
     return elysia
@@ -128,8 +157,8 @@ export class CrudApi {
           tags: [name],
           description: `Insert item to ${name}`,
           responses: {
-            '200': {
-              description: 'Successful response',
+            '201': {
+              description: 'Successful response. Item created',
             },
             '400': {
               description: 'Error response',
@@ -186,10 +215,11 @@ export class CrudApi {
       `/${name}/:id`,
       async ({ params, body, error }) => {
         try {
-          // @ts-ignore
           const item = await this.db
             .update(schema)
+            // @ts-ignore
             .set(body)
+            // @ts-ignore
             .returning({ updatedId: schema.id })
 
           return defaultItemResponse(item)
